requirements.txt
amqp==2.1.4
anyjson==0.3.3
appdirs==1.4.3
beautifulsoup4==4.6.0
billiard==3.5.0.2
boto==2.47.0
celery==4.0.2
Django==1.11.1
django-admin-sortable==2.1
django-appconf==1.0.2
django-cors-headers==2.0.2
django-embed-video==1.1.2
django-environ==0.4.3
django-filter==1.0.4
django-imagekit==4.0.1
django-ordered-model==1.4.1
django-solo==1.1.2
django-storages-redux==1.3.2
django-taggit==0.22.1
-e git://github.com/Arlefreak/django-taggit-serializer.git@1334108a8e7bc3d9cf0b4579aac146469783c371#egg=django_taggit_serializer
djangorestframework==3.6.3
greenlet==0.4.12
gunicorn==19.7.1
idna==2.5
kombu==4.0.2
Markdown==2.6.8
msgpack-python==0.4.8
neovim==0.1.13
olefile==0.44
packaging==16.8
pilkit==2.0
Pillow==4.1.1
psycopg2==2.7.1
pyparsing==2.2.0
pytz==2017.2
redis==2.10.5
requests==2.14.2
requests-file==1.4.2
six==1.10.0
tldextract==2.1.0
vine==1.1.3
